+++ packages/e2e/package.json
@@ -129,6 +129,7 @@
     "@babel/preset-env": "^7.18.2",
     "@babel/preset-typescript": "^7.17.12",
     "@cardano-sdk/dapp-connector": "^0.9.0",
+    "@cardano-sdk/tx-construction": "^0.4.0",
     "@cardano-sdk/projection": "^0.6.0",
     "@cardano-sdk/projection-typeorm": "^0.2.0",
     "@cardano-sdk/web-extension": "^0.9.0",
