+++ packages/tx-construction/package.json
@@ -65,6 +65,7 @@
     "@cardano-sdk/util": "^0.9.0",
     "lodash": "^4.17.21",
     "npm": "^9.3.0",
+    "rxjs": "^7.4.0",
     "ts-custom-error": "^3.2.0",
     "ts-log": "^2.2.4"
   },
