+++ packages/tx-construction/src/index.ts
@@ -1,2 +1,7 @@
+export * from './createTransactionInternals';
 export * from './fees';
 export * from './input-selection';
+export * from './output-validation';
+export * from './tx-builder';
+export * from './ensureValidityInterval';
+export * from './types';
