+++ packages/tx-construction/src/output-validation/index.ts
@@ -1,0 +1,2 @@
+export * from './outputValidator';
+export * from './types';
