+++ packages/tx-construction/src/tx-builder/index.ts
@@ -1,0 +1,4 @@
+export * from './OutputBuilder';
+export * from './finalizeTx';
+export * from './initializeTx';
+export * from './types';
