+++ packages/tx-construction/test/tsconfig.json
@@ -3,6 +3,11 @@
   "compilerOptions": {
     "baseUrl": "."
   },
+  "include": [
+    "./**/*.ts",
+    "../../core/test/mocks",
+    "../../input-selection/test/util"
+  ],
   "references": [
     {
       "path": "../src"
