--- packages/wallet/src/Transaction/index.ts
@@ -1,2 +1,0 @@
-export * from './createTransactionInternals';
-export * from './ensureValidityInterval';
