+++ packages/wallet/src/TxBuilder/index.ts
@@ -1,3 +1,1 @@
-export * from './types';
 export * from './buildTx';
-export * from './OutputBuilder';
