+++ packages/wallet/test/integration/CustomObservableWallet.test.ts
@@ -4,20 +4,16 @@
 import * as mocks from '../../../core/test/mocks';
 import { Cardano, Transaction } from '@cardano-sdk/core';
 import { GroupedAddress } from '@cardano-sdk/key-management';
+import { ObservableWallet, SingleAddressWallet, coldObservableProvider } from '../../src';
 import {
-  ObservableWallet,
   OutputValidator,
   ProtocolParametersRequiredByOutputValidator,
-  RewardAccount,
-  SingleAddressWallet,
-  StakeKeyStatus,
-  coldObservableProvider,
   createOutputValidator
-} from '../../src';
+} from '@cardano-sdk/tx-construction';
 import { RetryBackoffConfig } from 'backoff-rxjs';
 import { createStubStakePoolProvider } from '@cardano-sdk/util-dev';
+import { firstValueFrom, of, timer } from 'rxjs';
 import { dummyLogger as logger } from 'ts-log';
-import { of, timer } from 'rxjs';
 import { testAsyncKeyAgent } from '../../../key-management/test/mocks';
 import { usingAutoFree } from '@cardano-sdk/util';
 
@@ -75,10 +71,10 @@ describe('CustomObservableWallet', () => {
       // and another endpoint to get wallet's total reward accounts deposit
       const getAvailableRewardAccountsDeposit: () => Promise<Cardano.Lovelace> = async () => 200_000n;
       // and another endpoint to get wallet's reward accounts info (such as stake pool delegated to)
-      const getRewardAccountsDelegation: () => Promise<RewardAccount[]> = async () => [
+      const getRewardAccountsDelegation: () => Promise<Cardano.RewardAccountInfo[]> = async () => [
         {
           address: Cardano.RewardAccount('stake1u89sasnfyjtmgk8ydqfv3fdl52f36x3djedfnzfc9rkgzrcss5vgr'),
-          keyStatus: StakeKeyStatus.Unregistering,
+          keyStatus: Cardano.StakeKeyStatus.Unregistering,
           rewardBalance: 5000n
         }
       ];
@@ -143,17 +139,19 @@ describe('CustomObservableWallet', () => {
       it('can utilize SingleAddressWallet', () => {
         const wallet: SingleAddressWallet = {} as SingleAddressWallet;
         // this compiles
-        outputValidator = createOutputValidator(wallet);
+        outputValidator = createOutputValidator({
+          protocolParameters: () => firstValueFrom(wallet.protocolParameters$)
+        });
         outputValidator;
       });
 
       it('can provide an implementation that utilize SDK utils, but doesnt depend on full ObservableWallet', () => {
-        const protocolParameters$ = of<ProtocolParametersRequiredByOutputValidator>({
+        const protocolParameters: ProtocolParametersRequiredByOutputValidator = {
           coinsPerUtxoByte: 34_482,
           maxValueSize: 5000
-        });
+        };
         // this compiles
-        outputValidator = createOutputValidator({ protocolParameters$ });
+        outputValidator = createOutputValidator({ protocolParameters: async () => protocolParameters });
       });
 
       it('can provide custom implementation', () => {
