+++ packages/wallet/test/services/WalletUtil.test.ts
@@ -2,14 +2,12 @@ import * as Crypto from '@cardano-sdk/crypto';
 import * as mocks from '../../../core/test/mocks';
 import { AddressType, GroupedAddress } from '@cardano-sdk/key-management';
 import { CML, Cardano } from '@cardano-sdk/core';
+import { ProtocolParametersRequiredByOutputValidator } from '@cardano-sdk/tx-construction';
 import {
-  OutputValidator,
-  ProtocolParametersRequiredByOutputValidator,
   SingleAddressWallet,
   WalletUtilContext,
   createInputResolver,
   createLazyWalletUtil,
-  createOutputValidator,
   requiresForeignSignatures,
   setupWallet
 } from '../../src';
@@ -20,48 +18,6 @@ import { stakeKeyDerivationPath, testAsyncKeyAgent } from '../../../key-manageme
 import { waitForWalletStateSettle } from '../util';
 
 describe('WalletUtil', () => {
-  describe('createOutputValidator', () => {
-    let validator: OutputValidator;
-
-    beforeAll(() => {
-      validator = createOutputValidator({
-        protocolParameters$: of<ProtocolParametersRequiredByOutputValidator>({
-          coinsPerUtxoByte: 4310,
-          maxValueSize: 90
-        })
-      });
-    });
-
-    it('validateValue validates minimum coin quantity', async () => {
-      expect((await validator.validateValue({ coins: 2_000_000n })).coinMissing).toBe(0n);
-      expect((await validator.validateValue({ coins: 500_000n })).coinMissing).toBeGreaterThan(0n);
-    });
-
-    it('validateValue validates bundle size', async () => {
-      expect(
-        (
-          await validator.validateValue({
-            assets: new Map([
-              [Cardano.AssetId('b01fb3b8c3dd6b3705a5dc8bcd5a70759f70ad5d97a72005caeac3c652657675746f31333237'), 1n]
-            ]),
-            coins: 2_000_000n
-          })
-        ).tokenBundleSizeExceedsLimit
-      ).toBe(false);
-      expect(
-        (
-          await validator.validateValue({
-            assets: new Map([
-              [Cardano.AssetId('b01fb3b8c3dd6b3705a5dc8bcd5a70759f70ad5d97a72005caeac3c652657675746f31333237'), 1n],
-              [Cardano.AssetId('c01fb3b8c3dd6b3705a5dc8bcd5a70759f70ad5d97a72005caeac3c652657675746f31333237'), 2n]
-            ]),
-            coins: 2_000_000n
-          })
-        ).tokenBundleSizeExceedsLimit
-      ).toBe(true);
-    });
-  });
-
   describe('createInputResolver', () => {
     it('resolveInput resolves inputs from provided utxo set', async () => {
       const utxo: Cardano.Utxo[] = [
